Bug #1347: Add support for float quotas. The update quotas method needs to be changed too (cherry picked from commit 6cc05f5b81c7cdcf302d255a3d6490ea87921cc1)

+++ src/um/QuotaNetwork.cc
@@ -34,7 +34,7 @@ bool QuotaNetwork::check(Template * tmpl,  string& error)
     string net_id;
     int num;
 
-    map<string, int> net_request;
+    map<string, float> net_request;
 
     net_request.insert(make_pair("LEASES",1));
